+++ burst-kotlin-plugin-tests/src/test/kotlin/app/cash/burst/kotlin/TestInterceptorKotlinPluginTest.kt
@@ -1715,4 +1715,56 @@ class TestInterceptorKotlinPluginTest {
 
     assertThat(log).containsExactly("SampleInterface", "Rock", "SampleObject")
   }
+
+  @Test
+  fun testFunctionWithAnnotations() {
+    val log =
+      BurstTester(packageName = "com.example")
+        .compileAndRun(
+          SourceFile.kotlin(
+            "Main.kt",
+            $$"""
+        package com.example
+
+        import app.cash.burst.InterceptTest
+        import app.cash.burst.TestFunction
+        import app.cash.burst.TestInterceptor
+        import kotlin.test.Test
+
+        @Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS)
+        annotation class SampleAnnotation(val label: String)
+
+        @SampleAnnotation(label = "classy")
+        class SampleTest {
+          @InterceptTest
+          val interceptor = object : TestInterceptor {
+            override fun intercept(testFunction: TestFunction) {
+              log("function annotations: ${testFunction.functionAnnotations}")
+              log("class annotations: ${testFunction.classAnnotations}")
+
+              testFunction()
+            }
+          }
+
+          @Test
+          @SampleAnnotation(label = "funky")
+          fun happyPath() {
+            log("running happyPath")
+          }
+        }
+
+        fun main(vararg args: String) {
+          SampleTest().happyPath()
+        }
+        """,
+          )
+        )
+
+    assertThat(log)
+      .containsExactly(
+        $$"function annotations: [@org.junit.Test(expected=class org.junit.Test$None, timeout=0), @com.example.SampleAnnotation(label=funky)]",
+        "class annotations: [@com.example.SampleAnnotation(label=classy)]",
+        "running happyPath",
+      )
+  }
 }