+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [Unreleased]
 [Unreleased]: https://github.com/cashapp/burst/compare/2.11.0...HEAD
 
+Nothing yet!
+
+## [2.12.0-beta2]
+[2.12.0-beta2]: https://github.com/cashapp/burst/releases/2.12.0-beta2
+
 **Added**
 
 * Pass annotation instances to interceptors